requirements.txt
streamlit==1.28.0
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
protobuf==3.20.3
Pillow

